+++ dialogs/dialognormal.h
@@ -26,8 +26,8 @@ public:
     void             setTypeLine(const QString &value);
     QString          getFormula() const;
     void             setFormula(const QString &value);
-    qint32           getAngle() const;
-    void             setAngle(const qint32 &value);
+    qreal            getAngle() const;
+    void             setAngle(const qreal &value);
     qint64           getFirstPointId() const;
     void             setFirstPointId(const qint64 &value, const qint64 &id);
     qint64           getSecondPointId() const;
@@ -41,7 +41,7 @@ private:
     QString          pointName;
     QString          typeLine;
     QString          formula;
-    qint32           angle;
+    qreal            angle;
     qint64           firstPointId;
     qint64           secondPointId;
     DialogNormal(const DialogNormal &dialog);
