+++ tools/drawTools/vtoollinepoint.h
@@ -9,13 +9,13 @@ class VToolLinePoint : public VToolPoint
 public:
     VToolLinePoint(VDomDocument *doc, VContainer *data, const qint64 &id,
                    const QString &typeLine, const QString &formula,
-                   const qint64 &basePointId, const qint32 &angle, QGraphicsItem * parent = 0);
+                   const qint64 &basePointId, const qreal &angle, QGraphicsItem * parent = 0);
 public slots:
     virtual void      ChangedActivDraw(const QString newName); 
 protected:
     QString           typeLine;
     QString           formula;
-    qint32            angle;
+    qreal             angle;
     qint64            basePointId;
     QGraphicsLineItem *mainLine;
     virtual void      RefreshGeometry();
