+++ tools/modelingTools/vmodelinglinepoint.h
@@ -9,11 +9,11 @@ class VModelingLinePoint : public VModelingPoint
 public:
     VModelingLinePoint(VDomDocument *doc, VContainer *data, const qint64 &id,
                    const QString &typeLine, const QString &formula,
-                   const qint64 &basePointId, const qint32 &angle, QGraphicsItem * parent = 0);
+                   const qint64 &basePointId, const qreal &angle, QGraphicsItem * parent = 0);
 protected:
     QString           typeLine;
     QString           formula;
-    qint32            angle;
+    qreal             angle;
     qint64            basePointId;
     QGraphicsLineItem *mainLine;
     virtual void      RefreshGeometry();
